+++ core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/api/MybraryAuthApiModule.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.supabase
+package app.kaito_dogi.mybrary.core.supabase.api
 
 import app.kaito_dogi.mybrary.core.api.mybrary.MybraryAuthApi
 import dagger.Binds
